requirements.txt
click==6.7
feedparser==5.2.1
Flask==0.12.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
python-dotenv==0.8.2
PyYAML==3.12
Werkzeug==0.14.1
